implict enum to integer warning fix

+++ include/RE/A/ActorState.h
@@ -69,7 +69,7 @@ namespace RE
 		kWaitingForSitAnim = 2,
 
 		kIsSitting = 3,
-		kRidingMount = kIsSitting,
+		kRidingMount = static_cast<int>(kIsSitting),
 
 		kWantToStand = 4,
